bugfix: pid_t used as array index
In reserve_slot, a local variable `p` holds a process id, then is
re-used as an array index.  It's possible that the search for an empty
slot fails to find a valid slot, in which case `p` will contain a
process id, and the array index fetch will either segfault or
dereference an arbitrary position in memory.

The fix is to track the search loop with an integer index (not a pid_t),
initialized to a valid index (min_eternal_ix=0), and is always
valid to fetch.