fix read-overflow in gethost
The member utmpx.ut_line[] is not guaranteed to be null terminated.
The fix is to use strncmp instead of strcmp.

This is a preventative fix to address a compiler diagnostic,
rather than an actual bugfix. An overflow was unlikely
here because the ttyname would have had to exceeed the size of
ut_line, which is currently 32 chars.